fix: intからsize_tに暗黙的型変換が行われている箇所の修正 & CubeManager内でメンバ変数の変更を許容しない関数にconst修飾子を追加

+++ DX11Demo/src/Renderer.cpp
@@ -400,7 +400,7 @@ void Renderer::Render(float deltaTime)
     UINT offset = 0;
     ConstantBuffer cb;
 
-    for (int i = 0; i < g_cubeManager.size(); i++)
+    for (size_t i = 0; i < g_cubeManager.size(); i++)
     {
         Cube cube = g_cubeManager.getCube(i);
         XMMATRIX cubeWorld = XMMatrixScaling(cube.scale.x, cube.scale.y, cube.scale.z)
